+++ src/store/reducers/shareState.js
@@ -1,0 +1,10 @@
+import { UPDATE_SHARE_STATE } from "../actions/actionTypes";
+
+export default( state = {} , action ) => {
+	switch( action.type ) {
+		case UPDATE_SHARE_STATE:
+			return action.shareInfo;
+		default:
+			return state;
+	}
+}