+++ changelog.md
@@ -1,5 +1,14 @@
 #Change Log
 
+## 06/08/2019
+
+### Add Share Feature
+
+1.  Users can now share techniques or sequences via email
+
+#### Next major version will cause breaking changes
+1. Refactoring technique reference components to play nicely with updating techniques by using intermediate state to load form from
+
 ## 02/08/2019
 
 ### Change Log Creation